Added version 1.15.1147

+++ chromium/content/browser/shared_worker/shared_worker_host.cc
@@ -13,6 +13,7 @@
 #include "content/browser/shared_worker/shared_worker_content_settings_proxy_impl.h"
 #include "content/browser/shared_worker/shared_worker_instance.h"
 #include "content/browser/shared_worker/shared_worker_service_impl.h"
+#include "content/public/browser/browser_context.h"
 #include "content/public/browser/browser_thread.h"
 #include "content/public/browser/content_browser_client.h"
 #include "content/public/browser/render_process_host.h"
@@ -53,13 +54,13 @@ bool AllowIndexedDBOnIOThread(const GURL& url,
 }  // namespace
 
 SharedWorkerHost::SharedWorkerHost(
+    SharedWorkerServiceImpl* service,
     std::unique_ptr<SharedWorkerInstance> instance,
-    int process_id,
-    int route_id)
+    int process_id)
     : binding_(this),
+      service_(service),
       instance_(std::move(instance)),
       process_id_(process_id),
-      route_id_(route_id),
       next_connection_request_id_(1),
       creation_time_(base::TimeTicks::Now()),
       interface_provider_binding_(this),
@@ -70,14 +71,14 @@ SharedWorkerHost::SharedWorkerHost(
 SharedWorkerHost::~SharedWorkerHost() {
   UMA_HISTOGRAM_LONG_TIMES("SharedWorker.TimeToDeleted",
                            base::TimeTicks::Now() - creation_time_);
-  if (!closed_ && !termination_message_sent_) {
-    SharedWorkerDevToolsManager::GetInstance()->WorkerDestroyed(process_id_,
-                                                                route_id_);
-  }
+  if (!closed_ && !termination_message_sent_)
+    SharedWorkerDevToolsManager::GetInstance()->WorkerDestroyed(this);
 }
 
-void SharedWorkerHost::Start(mojom::SharedWorkerFactoryPtr factory,
-                             bool pause_on_start) {
+void SharedWorkerHost::Start(
+    mojom::SharedWorkerFactoryPtr factory,
+    bool pause_on_start,
+    const base::UnguessableToken& devtools_worker_token) {
   blink::mojom::WorkerContentSettingsProxyPtr content_settings;
   content_settings_ = std::make_unique<SharedWorkerContentSettingsProxyImpl>(
       instance_->url(), this, mojo::MakeRequest(&content_settings));
@@ -96,9 +97,9 @@ void SharedWorkerHost::Start(mojom::SharedWorkerFactoryPtr factory,
       instance_->creation_address_space()));
 
   factory->CreateSharedWorker(
-      std::move(info), pause_on_start, instance_->devtools_worker_token(),
-      route_id_, std::move(content_settings), std::move(host),
-      mojo::MakeRequest(&worker_), std::move(interface_provider));
+      std::move(info), pause_on_start, devtools_worker_token,
+      std::move(content_settings), std::move(host), mojo::MakeRequest(&worker_),
+      std::move(interface_provider));
 
   // Monitor the lifetime of the worker.
   worker_.set_connection_error_handler(base::BindOnce(
@@ -111,7 +112,9 @@ void SharedWorkerHost::AllowFileSystem(
   BrowserThread::PostTask(
       BrowserThread::IO, FROM_HERE,
       base::BindOnce(&AllowFileSystemOnIOThread, url,
-                     instance_->resource_context(),
+                     RenderProcessHost::FromID(process_id_)
+                         ->GetBrowserContext()
+                         ->GetResourceContext(),
                      GetRenderFrameIDsForWorker(), std::move(callback)));
 }
 
@@ -121,17 +124,20 @@ void SharedWorkerHost::AllowIndexedDB(const GURL& url,
   BrowserThread::PostTaskAndReplyWithResult(
       BrowserThread::IO, FROM_HERE,
       base::BindOnce(&AllowIndexedDBOnIOThread, url, name,
-                     instance_->resource_context(),
+                     RenderProcessHost::FromID(process_id_)
+                         ->GetBrowserContext()
+                         ->GetResourceContext(),
                      GetRenderFrameIDsForWorker()),
       std::move(callback));
 }
 
 void SharedWorkerHost::TerminateWorker() {
+  // This can be called twice in tests while cleaning up all the workers.
+  if (termination_message_sent_)
+    return;
   termination_message_sent_ = true;
-  if (!closed_) {
-    SharedWorkerDevToolsManager::GetInstance()->WorkerDestroyed(process_id_,
-                                                                route_id_);
-  }
+  if (!closed_)
+    SharedWorkerDevToolsManager::GetInstance()->WorkerDestroyed(this);
   worker_->Terminate();
   // Now, we wait to observe OnWorkerConnectionLost.
 }
@@ -164,15 +170,13 @@ void SharedWorkerHost::OnContextClosed() {
   // being sent to the worker (messages can still be sent from the worker,
   // for exception reporting, etc).
   closed_ = true;
-  if (!termination_message_sent_) {
-    SharedWorkerDevToolsManager::GetInstance()->WorkerDestroyed(process_id_,
-                                                                route_id_);
-  }
+  if (!termination_message_sent_)
+    SharedWorkerDevToolsManager::GetInstance()->WorkerDestroyed(this);
 }
 
 void SharedWorkerHost::OnReadyForInspection() {
-  SharedWorkerDevToolsManager::GetInstance()->WorkerReadyForInspection(
-      process_id_, route_id_);
+  if (!closed_ && !termination_message_sent_)
+    SharedWorkerDevToolsManager::GetInstance()->WorkerReadyForInspection(this);
 }
 
 void SharedWorkerHost::OnScriptLoaded() {
@@ -235,6 +239,11 @@ bool SharedWorkerHost::ServesExternalClient() {
   return false;
 }
 
+void SharedWorkerHost::BindDevToolsAgent(
+    blink::mojom::DevToolsAgentAssociatedRequest request) {
+  worker_->BindDevToolsAgent(std::move(request));
+}
+
 void SharedWorkerHost::OnClientConnectionLost() {
   // We'll get a notification for each dropped connection.
   for (auto it = clients_.begin(); it != clients_.end(); ++it) {
@@ -251,8 +260,7 @@ void SharedWorkerHost::OnClientConnectionLost() {
 void SharedWorkerHost::OnWorkerConnectionLost() {
   // This will destroy |this| resulting in client's observing their mojo
   // connection being dropped.
-  static_cast<SharedWorkerServiceImpl*>(SharedWorkerService::GetInstance())
-      ->DestroyHost(process_id_, route_id_);
+  service_->DestroyHost(this);
 }
 
 void SharedWorkerHost::GetInterface(
